refactor(router): move the task helpers beside the methods that use them
Phase 2 of #1256. `router.rs` goes 4636 to 4454 lines; the nine free
functions join `router/task_ops.rs`.

They sat at the top of `router.rs`, above the type they serve, and their only
callers are the task methods extracted in #1319: `task_store_error`,
`discard_unprepared_task`, both `client_declares_tasks` variants,
`decode_input_responses`, `expired_task_error`, `unknown_task_error`,
`tasks_client_capabilities`, and `validate_input_required_result`.

Seven are `pub(super)` so `router.rs` keeps calling them unqualified through
a `use`. `tasks_client_capabilities` was already `pub(crate)` and stays that
way. `expired_task_error` and `unknown_task_error` turned out to have no
caller outside the task group, so they are private to it now, which is
tighter than before the move.

The default-feature build caught the one real hazard here.
`validate_input_required_result` is `#[cfg(feature = "stateless")]`, and an
unconditional `use` of it compiles fine under `--all-features` while breaking
every build without that feature. The import is gated to match.

1030 lib tests before and after, and the build is clean on default,
all-features, and four intermediate feature sets.

Refs #1256
